% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.random.R
\name{generate.random}
\alias{generate.random}
\title{Generate used and random points}
\usage{
generate.random(
  n.gps = nrow(gps.s),
  random = dp.s,
  n.samp = 10,
  max.dist = 60000,
  dist.raster = short.path,
  qweights = "inverse.dist"
)
}
\arguments{
\item{n.gps}{Number of rows in gps table}

\item{random}{Spatialpointsdataframe created by st_as_sf of random points}

\item{n.samp}{The number of random points to select for used GPS/used location}

\item{max.dist}{The maximum distance in which to select the random (and used points). So also the used points > max.dist are removed.}

\item{dist.raster}{Distance raster}

\item{qweights}{Type of extraction to do}
}
\value{
A SpatialPointDataFrame with the used and available locations
}
\description{
Generate used and random points
}
\examples{
generate.random2()
}
