% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbathymetry.R
\name{getbathymetry}
\alias{getbathymetry}
\title{Extract bathymetry from Emodnet:}
\usage{
getbathymetry(
  name = "emodnet:mean",
  crs.epsg = "EPSG:4326",
  resolution = "0.2km",
  xmin = 15,
  xmax = 20.5,
  ymin = 30,
  ymax = 32.5
)
}
\arguments{
\item{name}{E.g. "emodnet:mean"}

\item{crs.epsg}{crs code, e.g. "EPSG:4326"}

\item{resolution}{E.g. "0.2km"}

\item{xmin}{xmin of box in longitude}

\item{xmax}{xmax of box in longitude}

\item{ymin}{ymin of box in latitude}

\item{ymax}{ymax of box in latitude}
}
\value{
An object of type raster containing bathymetry values
}
\description{
Extract bathymetry from Emodnet:
}
\examples{
getbathymetry()
}
