% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.extract.R
\name{nn.extract}
\alias{nn.extract}
\title{Nearest neighbour extraction of raster values}
\usage{
nn.extract(the.raster, pp = as.data.frame(dp.s), max.dist = 5000)
}
\arguments{
\item{the.raster}{Raster object for the environmental variable. Should be same projection as pp}

\item{pp}{A data.frame (with coordinates in first two columns), a SpatialPointsDataFrame or sf for which to extract the environmental values from raster}

\item{max.dist}{The maximum distance for the nearest neighbour.}
}
\value{
pp with the added columns value: the value of the extracted environmental value for the raster and dist: the distance to the nearest point in the raster. dist = 0 if the point fals within a cell.
}
\description{
Nearest neighbour extraction of raster values
}
\examples{
nn.extract()
}
