% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.clustering.R
\name{point.clustering}
\alias{point.clustering}
\title{Title}
\usage{
point.clustering(pp = haulout, d = 1000)
}
\arguments{
\item{pp}{2-column matrix/data.frame with x and y coordinates to be clustered. Coordinates should by X and Y in equidistant projection (e.g. utm)}

\item{d}{Distance; same units as coordinates X and Y}
}
\value{
A list with clust: the cluster id and cent: the central point of each cluster
}
\description{
Title
}
