% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile.to.raster.points.R
\name{shapefile.to.raster.points}
\alias{shapefile.to.raster.points}
\title{Convert land shapefile to raster}
\usage{
shapefile.to.raster.points(
  shapefile = land,
  res = 501,
  the.extent = c(2, 8, 51, 56)
)
}
\arguments{
\item{shapefile}{A shapefile to convert to raster}

\item{res}{The resolution of the output raster}

\item{the.extent}{Boundaries of the raster. A vector of size 4 with x1,x2,y1,y2}
}
\value{
A data.frame
}
\description{
Convert land shapefile to raster
}
\examples{
shapefile.to.raster.points()
}
