% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortest.path.dist.R
\name{shortest.path.dist}
\alias{shortest.path.dist}
\title{Calculate shortest path}
\usage{
shortest.path.dist(
  graph.list = g2.list,
  the.raster = depth,
  start.x = 550000,
  start.y = 5900000,
  max.radius = 20000
)
}
\arguments{
\item{graph.list}{A graph list object created by dist.graph, see help(dist.graph)}

\item{the.raster}{A object of type raster used to define the resolution of the output grid}

\item{start.x}{The x-coordinate of the starting point}

\item{start.y}{The y-coordinate of the starting point}
}
\value{
The point grid with the shortest path distance
}
\description{
Calculate shortest path
}
\examples{
shortest.past.dist()
}
